Moved service class, integrate first round of review feedback
