+++ openeo/rest/connection.py
@@ -31,8 +31,8 @@ from openeo.rest.datacube import DataCube
 from openeo.rest.imagecollectionclient import ImageCollectionClient
 from openeo.rest.job import RESTJob
 from openeo.rest.rest_capabilities import RESTCapabilities
+from openeo.rest.service import Service
 from openeo.rest.udp import RESTUserDefinedProcess, Parameter
-from openeo.service import Service
 from openeo.util import ensure_list, legacy_alias, dict_no_none, rfc3339
 
 _log = logging.getLogger(__name__)
@@ -799,7 +799,7 @@ class Connection(RestApiConnection):
         service_id = response.headers.get("OpenEO-Identifier")
         return Service(service_id, self)
 
-    @deprecated("Use :py:meth:`openeo.service.Service.delete_service` instead.", version="0.7.0")
+    @deprecated("Use :py:meth:`openeo.rest.service.Service.delete_service` instead.", version="0.7.1")
     def remove_service(self, service_id: str):
         """
         Stop and remove a secondary web service.
